requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.2
python-dotenv==1.0.0
pydantic-settings==2.1.0
PyJWT==2.8.0
python-multipart==0.0.6
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.7

